테스트 스킬 코드 추가

+++ Source/PalworldZA/Character/Pokemon/TestPokemon.cpp
@@ -20,7 +20,7 @@ void ATestPokemon::Tick(float DeltaTime)
 	Super::Tick(DeltaTime);
 }
 
-bool ATestPokemon::UsingSkill(int SkillNumber)
+bool ATestPokemon::UsingSkill(uint8 SkillNumber)
 {
 	bool Value = Super::UsingSkill(SkillNumber);
